SMBFS: fix detecting if a volume is in-use
The SMBFS driver along with the remotefs module rely on the volume
status field when checking if it's attached.

While in most cases this is fine, it will cause issues when backing
up volumes. In this situation, the volume status will be 'backing-up',
in which case the driver will wrongfully consider that the volume is
detached (by not having the 'in-use' state).

In particular, this affects the volume snapshot workflow.

This change fixes this check, adding a helper method which detects
the volume attach status by looking at the volume 'attach_status'
field.

Change-Id: I154d85cde693c6e77e22a677cae7150be48fce07
Partial-Implements: windows-smb-backup