VNX: fix options incorrect behavior
Since Newton, the following options were incorrectly handled:

max_luns_per_storage_group: this option was ignored, and implicitly set
to 255 in storops, thus there is not way to set it to a larger number
for admin. This commit will add this feature back.

check_max_pool_luns_threshold: this option was obsoleted since driver
can detect this value automatically. This commit will remove this
option for driver code.

Closes-bug: #1738681

Change-Id: If3ca0999a745e6977345f208fc6edd00c86ccea9