Enable fail back in SolidFire driver
This change fixes up a few things in the SolidFire driver to enable the
ability to "fail back" to the original cluster in a replication scenario.

We're still assuming a cluster-wide fail over, but now an admin has the
ability to specify subsequent `cinder failover-host` calls to switch
back and forth between their SF clusters.

This change promotes the target on a fail over and also now attempts to
set up replication back to the original src  as well (assuming the
original is available and not a puddle of melted goo).  This means that
writes to the new back end will be replicated back to the src, so if a
user so chooses they can fail back to the original and have their new
data.

Change-Id: I678986e557755c2fe4183927c17342e430f5df0d