Add admin ctxt to request qos_specs in SolidFire
We used to be able to call qos_specs_get without admin context,
however something has changed and this now fails at the object
layer in base:get_by_id.  This means that now if one uses QoS-Specs
with SolidFire and attempts to retype to another QoS Spec type,
the call will fail with no response data due to a policy/permissions
error.

It turns out that the qos_specs.py api call will let you either
pass in a context or pass in None and it implicitly gets an admin
context.  I prefer explicit behaviors, so add an explicit get admin
context in the SF driver for this call.

Change-Id: Idf10bfb8f5e738d7698233b7bdeaf45bef8914e2
Closes-Bug: #1738800