Update tgt iSCSI target on volume extend
When extending an in-use volume, the volume will be extended but
the target exposing the volume will still report the original size.

The target needs to be updated to pick up the change, but since it
is in-use the update must be forced. This can cause IO disruption
if there is active IO to the target, however retries will cause the
next attempt to succeed and there will just be a few second IO
pause on the host while the target session is reestablished. Not
ideal, but it does recover with minimal impact on the host.

And operators probably shouldn't extend volumes without pausing
active IO to the volume anyway.

See bug report comments for more testing details. Marking this as
partial as it only addresses the tgt target and further
investigation is needed to see if the other target drivers need
similar updates.

Change-Id: I0f7b58751c6d8d05cad5b1b95c4e21c8b5d36692
Partial-bug: #1731984