StorageCenter: Fix volume mapping for API v3.1
Volume attachment fails on StorageCenter API v3.1 because the backend
instead of returning wwn or WWN as the key in the controller port it
returns wWN, thus preventing the driver from finding the mapping.

This patch adds support to any combination of upper and lower case of
the WWN key.

Closes-Bug: #1739651
Change-Id: Ie6ec59b4f6d31cb72e37f25f5347c815fee5af48