Do not explicitly list thin provisioning tools for RPM
Ubuntu requires an additional package to be installed to enable thin
provisioning, but Red Hat based platforms have this included with the
base LVM2 package. An explicit item was added to bindep for the RPM
package that includes these thin tools, but that is not needed and
should be removed.

Change-Id: I7e7575746911e8d19a3804fa08b357c9563713f1