Enable multiattach capability for lvm driver
Multiattach was globally disabled for all in-tree drivers
in e499ab995de9c6dca353a115c6c68bd04594cdd9 but we'll need
to start enabling these to support the functionality.

A policy rule in the API, and/or an extra spec on the volume
type, should be used to distinguish which backends, or the
deployment in general, can support multiattach.

Related to blueprint multi-attach-v3-attach

Change-Id: I158c6f20e3e6a24bd2e5299abbeb3fc5208e5885