Handle InvalidVolume when creating a volume attachment
If you try to attach a non-multiattach volume to more than
once instance using the 3.27 volume attachments API, the volume
API would return a 500 response because the volume is in-use.

This translates to a 500 in the compute API because nova isn't
expecting a random client exception.

Treat this like how os-reserve could fail in the same situation
and handle InvalidVolume which will be converted to an HTTPBadRequest
by ResourceExceptionHandler elsewhere in the WSGi stack.

Change-Id: I1a8068043159b034d700419675af87fc17e1faf5
Closes-Bug: #1741112