VMware:Config option to disable lazy volume create
Currently the backend volume is created lazily for raw
volume create operations. The backend volume is created
during the initial attach. This is to minimize unnecessary
volume migrations across datastores in the backend which
may be required if the backend volume is not accessible to
instance VM's ESX host. In certain vCenter configurations,
it may be required to disable this lazy creation. This would
reduce the time spent in the backend during volume attach.
Introducing a config option to disable lazy volume create
if necessary.

Change-Id: I3feb6b382ff75eb53d53f145ae3773b6320e703c
Closes-bug: #1739390