Tests: Fix XIV test failure with hash randomization
_get_fc_targets creates a set() from target_wwpns.
It then sorts the contents of that set using a comparison that
only checks the last digit of the entry.

This is not a stable sort, so the order of the list output
from _get_fc_targets is non-deterministic.

This appears to be fine in practice, so just fix the unit test.

Needed-By: I4c7396f9d8e58ac8420754503f8ed60cd7473426
Change-Id: I281ad59fa518f936aa8d1b518acf61800ba949f5