Improve multiattach error message
With multi-attach on the way, the current message when you try to attach
a non-multiattach enabled disk is less than perfect. If you've forgotten
to set multiattach on volume create, you will get this error when you
try to attach it to the second instance:

"volume is already attached"

This updates the error message to make it clearer.

Closes-Bug: 1741943

Change-Id: I7110cb87388ec8d39aa8b604208a161e890f3ab2