3PAR: fix create group from source functionality
when we create a group from source having tiramisu enabled
on group, adding volumes from new group to 3par remote
copy group is giving an error.
This is because of 3par has limitation that physical copy
can not be added to remote copy group.
in create group from source function, we are making use
of online copy feature of 3par for copying of data from
source volume to destination volume, which creates
physical copy of destination volume till completion of
online copy due to which addition of such physical copy
to remote copy group is failed.

To overcome this, we will wait till online copy completes
in 3par for volumes. Because after completion of online copy,
destination volumes becomes normal virtual volume in 3par
and we can easily add such volumes to remote copy group.

Change-Id: Idc74cc68350833e3c9ac85f36c8f058beeca6bc9
Closes-Bug: #1742092