api-ref: fix the response parameter for multiattach
The multiattach parameter is optional when creating
a volume but is required to be in the response when
showing a volume.

This fixes the parameter usage, renames the parameters
so they are more clear (multiattach_1 was for the response
but was never used), and also adds a note to the multiattach
request parameter about how support is dependent on the volume
type being used to create the volume.

Change-Id: Ifaa29140a8e8a9b8090f33fb83e4bb5c98ee847f