Fix combination of parameters for update APIs
For group_type and volume_type update APIs, If user passes
description or is_public parameter in the request body then name
can be null. But after schema validation [1][2], it is failing with
BadRequest.

This patch allows users to pass 'name' as None if "description"
or is_public" parameter is passed in the request body.

[1]: https://review.openstack.org/#/c/520561/
[2]: https://review.openstack.org/#/c/519643/

Change-Id: Id5d5f5ed8b449c9f1059528de241417f231f90b5
Closes-Bug: #1742940