Storwize: Use passed volume names in migrate
For the storwize driver, update_migrated_volume() impl, the
original and temporary volume names were generated from a
configuration template value, rather than using the names of
the volumes passed. This only works if the storwize driver
created the original volume, but breaks down if it originated
elsewhere.
Make a small change to use the passed in volume name properties
instead. Add a couple asserts to the test cases.

Change-Id: I1c719e6858426416b56357d2f071584209757a78
Closes-Bug: #1721583