Fix: Propagate OS global request ID
Cinder doesn't propagate global request IDs internally, so any external
request that comes with it will have it logged at the API service and
then we'll lose this information in all the other services.

We are properly serializing this information on our RPC calls, but we
are not reconstructing the Context properly on the receiving end.

This patch ensures that the "global_request_id" field from the context
is properly added on RPC reception.

Closes-Bug: #1743462
Change-Id: I35834a43298b5287d5cd3c37baa82903bdffda12
(cherry picked from commit 28ada504af4084f72f3503d77806075e7ad509b0)