Fix compression related tests
Tests for chunked backup drivers are using a randomly generated file to
test the compression, which means that in most cases due to the
randomization the generated file is bigger than the original one, so we
end up doing a non compressed backup, which defeats the purpose of the
tests.  More so when you want to test the decompression, since the
backup won't be compressed and there won't be any decompression.

To fix this we generate, for those cases, a simple file that can always
be effectively compressed.

Change-Id: Idbfeeb2604bc84dc0a17c633b782000d94665bbe
(cherry picked from commit 8f3dfd7299a46be106d5b571c442a79986d66883)