Fix cinder quota-usage error
When backing up a snapshot or 'in-use' volume,
cinder creates a temp volume from the snapshot
or 'in-use' volume, then backup the temp volume.
The temp volume doesn't use the volume-quota,
but deleting the temp volume when finishing
backup will minus quota. This process leads to
a wrong quota-usage. So add admin volume metadata
when creating the temporary volume and then use
that data to detect temporary volume. Then skip
handling quota when deleting the temp volume.

Change-Id: If100a678cf9062c4078f850d8edc001b6b0705f4
Closes-Bug: #1670636