Use provider_id for SolidFire Volume lookups
Over the years the various iterations of volume-get in the SolidFire
driver has gotten a bit redundant.  The biggest problem with all of
these changes comes in when we start dealing with things like managed
volumes, most of the methods we have in place expect the volume-name
on the SolidFire backend to match the cinder UUID.

In the case of manage, migrate and some other cases however, this
isn't a valid assumption.  SolidFire doesn't allow changing the name
of an existing volume.  You're only choice in that case is to clone
or copy to a new volume, but that's terribly heavy weight in this case.

So, what we need to do is harden the volume_get method a bit.  We
want to do a better job here, first by attempting to use the provider_id
info that's stored with a volume, from there we can try some failback
methods like looking for name:uuid map and finally checking attributes.

This should make things much more efficient for delete calls, and fix
the current bug due to imported volumes not having the cinder-uuid as
the name.

We need to do some follow up work and consolidate all of these get calls,
but for now we just fix the recently reported bug.

Change-Id: I1a1d02a1407926c22fbadcc1049434e24b92549b
Closes-Bug: #1744005