Netapp Ontap: Adds support for auto-max-over-subscription
After the addition of the auto-max-over-subscription on Cinder, drivers that
somehow used the CONF.max_over_subscription_ratio in the driver where marked as
not supporting the feature. As Netapp Ontap drivers used the option to do some
calculations, they were marked as so.

In this patch we remove the function that did this calculation as the function
(_share_has_space_for_clone() is not actually necessary as currectly all volume
clones are checked by the scheduler.)

Change-Id: I25713b402ad93e0b4aa1861b4ccd9e05ee496200
Depends-on: If30bb6276f58532c0f78ac268544a8804008770e