Add back support for the multiattach flag for volume create
As the 'multiattach' request parameter in volume create is
not formally deprecated out of the REST API via microversion,
we need to keep the functionality working.

Currently the 'multiattach' attribute on the volume is set by
the volume_type only, which means that we cannot create a
multi-attach volume the old way, the volume gets created with
'multiattach': false.

This patch fixes that.

Closes-Bug: #1745219

Change-Id: Iac67f112b0dc9353c6a66e6fbc81cc8324a2b37c