NetApp: Add use-exact-size parameter when creating a LUN on ONTAP iSCSI
When a cinder volume is created the ontap backend fits the LUN to the
best geometry. It makes the image uploaded from a volume larger than
expected and a volume of larger size must be created from that image.
Using the use-exact-size parameter the backend does not fit to best
geometry, the image uploaded from that volume has the same size and a
volume of the same size may be created from that image.

Nevertheless this parameter is available only in Data ONTAP 9.1 (ontapi
version 1.110) and later.

Closes-Bug: #1731474
Change-Id: I0e21cbcb6effa1e72999580564099976511ca4a9
(cherry picked from commit 67391f1f0f30172190882e7d3f4a4ddc271dfa00)