[1/2] Fix mock for sleep mocks in volume drivers tests
According to `Applying the same patch to every test method` section in
the [1] we shouldn't use mock decorators on setUp method. This patch fixes it
moving mocks into the setUp methods.

This is the first patch in the series. We need to fix the sililar issue in IBM
drivers but this requires more work to fix them. So let's do fix this issue
step-by-step.

[1] https://docs.python.org/3.5/library/unittest.mock-examples.html

Change-Id: I4c2097307f812eb5847367b58ef6ba65c23b3a26