Simplify keystone password assertions
With an update to keystoneauth1 > 3.3.0, some of the arguments
to the password plugin change. Some of our unit tests were a
little too zealous in checking the expected parameters that get
passed in to this.

Since we just need to verify that the password plugin is being
used, and we don't necessarily care about each of the values
being passed, just assert that the password plugin was called,
check some specific values, and keep it simple.

Change-Id: Id0d3b06b2debad70d36c253a3c1ee9d561a22e64