Remove logging on Swift backup obj writer
Since change I1f1d9c0d6e3f04f1ecd5ef7c5d813005ee116409 we are running
parts of the backups on native threads, which due to an eventlet bug [1]
have bad interactions with greenthreads, so we have to avoid any logging
when executing code in a native thread.

This patch removes the MD5 logging on the SwiftObjectWriter close
method and adds comments and docstring referring to this limitation.

[1] https://github.com/eventlet/eventlet/issues/432

Closes-Bug: #1745168
Change-Id: I0857cecd7d8ab0ee7e3e9bd6e15f4987ede4d653