Add debug logging around multiattach status changes
When deleting an attachment, we now need to check for any other
attachments for the volume and try to set the status correctly
based on any of those remaining attachments.

There's not very much logging around this to help us out if
something goes wrong, so this patch just proposes some debug
level logging be added around the process.

Change-Id: I64b3046cfb3355b547e158783131338b8b0180d1