api-ref: update migration_policy retype note about encrypted volumes
Nova has historically not supported swap volume operations (via
cinder volume retype / volume migration) for an in-use encrypted
volume.

That was recently addressed via nova bug 1739593.

However, as of change Ibfa64f18bbd2fb70db7791330ed1a64fe61c1355
in nova, depending on the version of libvirt/qemu on the compute host,
a luks-encrypted volume will use native qemu luks decryption. That
does not yet support blockRebase operations which are used during
swap volume due to https://bugzilla.redhat.com/show_bug.cgi?id=760547.

So it's safe to say that for now, a retype which involves a volume
migration is not supported for an in-use encrypted volume.

Change-Id: I7ce992f51d50d00950d3fc4ebb44b69a31a94787
Related-Bug: #1739593