[Trivial] SMBFS: fix provisioning type usage
At the moment, the SMBFS driver does not honor the configured
provisioning type.

The issue is that we're checking CONF.backend_defaults instead
of the driver configuration object, for which reason if the option
is defined in the according backend config group, it will be ignored.

This change fixes this issue, updating the config option checks.

Change-Id: Ieb9e9edb1bc46d8b6ca31480d963f9dc9c2331de
Closes-Bug: #1745602