3PAR: fix cloning issue in create group from source.
When group or group snapshot is created from two
different volume types having provisioning set
one volume type and compression set on other
volume type and then if we create group from
existing group or group snapshot then we are
making use of source snapshot and online copy
feature of 3par for copying data from source
to destination.

But we have not kept mapping of source volume and
its snapshot and because of this, from any snapshot
of data gets copied to any destination volume and
created issues mentioned in bug.

This patch creates mapping of source volume and its
Snapshot, so correct snapshot get selected for correct
destination volume for copying of data.

Change-Id: I435cd47908fefaa767063fa4c61f3a4d76ffc4ae
Closes-Bug: #1744025