Fix leftovers after backup abort
When aborting a backup on any chunked driver we will be leaving chunks
in the backend without Cinder knowing so and with no way of deleting
them from Cinder.  In this case the only way to delete them is going to
the storage itself and deleting them manually.

Another issue that will happen if we are using a temporary resource for
the backup, be it a volume or a snapshot, is that it will not be cleaned
up and will be left for us to manually issue the delete through the
Cinder API.

The first issue is caused by the chunked driver's assumption that the
`refresh` method in an OVO will ignore the context's `read_deleted`
configuration and always read the record, which is not true.  And since
it doesn't work when the record is deleted there will be leftovers if
the status of the backup transitions to deleted during the processing of
a chunk.

The second issue is caused by the same thing, but in this case is when
the backup manager refreshes the backup OVO to know the temporary
resource it needs to clean up.

This patches fixes the incorrect behavior of the backup abort mechanism
to prevent leaving things behind.

Closes-Bug: #1746559
Change-Id: Idcfdbf815f404982d26618710a291054f19be736
(cherry picked from commit 4ff9e63707e2c4cf5869f28e3e86fd0606d2db9a)
(cherry picked from commit d65444ce7df88292581d0726f9eb633be2292287)
Conflicts:
	cinder/backup/chunkeddriver.py
(cherry picked from commit 94393daaa057b604ab212a99cd5cd18c693c95c1)
Conflicts:
	cinder/backup/manager.py
	cinder/tests/unit/backup/test_backup.py