DS8K: correct in-use replication vol status
Correct in-use replication volume status after failover/failback.
Currently the status of in-use replication volume is changed to
available after failover in ds8k cinder driver to make sure the
replication volume can be attached again after failover. So does
failback operation. However, nova is using the new cinder attach
api right now. The in-use volume can not be attached to the same
instance again even the volume status is changed to available and
attach_status is changed to detached in cinder driver by force.
This is due to nova stores the combined relationship of volume_id
and instance.uuid in nova db as BlockDeviceMapping. This is to avoid
attaching the same volume to the same instance twice. So the
attaching request can not be sent to cinder if there is attached
record in BlockDeviceMapping between the volume and instance. So this
patch keeps replication volume status during failover and faiback.

Change-Id: Ia66d3522bbe107fb7ab077ab948393c2b7f34737
Closes-Bug: 1746732