Make CinderException format errors fatal for tests
This patch borrows very heavily from Nova change
I680fd46d029ff58bd3b72ef7c7903c2271b26549.

We have a config option to treat exception message formatting errors
as fatal, but it turned out our overriding of that config option was
not actually working. Since this is really something only useful for
testing, and we pretty much always want errors like that caught in
testing, this deprecates the option and changes the code to always
enforce raising errors during unit testing when there are problems
formatting any exception messages.

The previous config option is pretty much useless now, but this way
we are following the deprecation policy and we can remove it right
away in Rocky.

Closes-Bug: #1728984
Change-Id: Iaeb2657a6622cede32c5a59327393ca20758f3c3