RBD: Don't query Ceph on stats for exclusive pools
Collecting stats for provisioned_capacity_gb takes a long time since we
have to query each individual image for the provisioned size.  If we are
using the pool just for Cinder and/or are willing to accept a potential
deviation in Cinder stats we could just not retrieve this information
and calculate this based on the DB information for the volumes.

This patch adds configuration option `rbd_exclusive_cinder_pool` that
allows us to disable the size collection and thus improve the stats
reporting speed.

Change-Id: I32c7746fa9149bce6cdec96ee9aa87b303de4271
Closes-Bug: #1704106
(cherry picked from commit f33baccc3544cbda6cd5908328a56096046657ed)